requirements.txt
atlassian-python-api
icalendar